feat: Add list of recently active followed users
For each user you are following, there is a function to display a list of recent posts (the last 3 posts). Users who have not posted in a while will be prioritized and displayed at the top of the list.
This feature avoids having the posts of some users who post frequently drown out the posts of those who post at a slower pace.