fix: fix default value of note count public range (public -> private)
Existing migration file (1702718871542-notesvisibility.js) and
UserProfile.ts, the default value of notesVisibility was set to 'private'.
but in the actual database it was set to 'public'.

The number of notes for new users is now automatically set to private,
A migration has been added to explicitly correct the default value in the database so that the number of notes for new users is automatically set to private.